Bundle declaration files into just one top-level index.d.ts for the whole library (plus one for the validation sub-library). Add ProtocolHttpModule to remove need for dependents of Protocol Common to import the proper nest axios version. Rename http-context -> http to better describe its contents. Fix some grammar errors in documentation.
