+++ dts.json
@@ -1,0 +1,16 @@
+{
+  "compilationOptions": {
+    "preferredConfigPath": "./tsconfig.json"
+  },
+
+  "entries": [
+    {
+      "filePath": "./src/index.ts",
+      "outFile": "./dist/index.d.ts"
+    },
+    {
+      "filePath": "./src/validation/index.ts",
+      "outFile": "./dist/validation/index.d.ts"
+    }
+  ]
+}
