+++ src/index.ts
@@ -1,9 +1,4 @@
 // entry point for the npm package
-
-export const version = (): string => {
-    return '1.0';
-};
-
 export * from './config.module.js';
 export * from './constants.js';
 export * from './datadog.logger.js';
@@ -18,10 +13,15 @@ export * from './protocol.error.interface.js';
 export * from './protocol.errorcode.js';
 export * from './protocol.exception.filter.js';
 export * from './protocol.exception.js';
-export * from './protocol.http.service.js';
 export * from './protocol.utility.js';
 export * from './random.string.generator.js';
 export * from './security.utility.js';
 export * from './trace.decorator.js';
 export * from './tracer.js';
-export * from './http-context/index.js';
+export * from './http/http.constants.js';
+export * from './http/protocol.http.service.js';
+export * from './http/protocol.http.module.js';
+export * from './http/request.context.middleware.js';
+export * from './http/request.context.module.js';
+export * from './http/request.context.js';
+
