--- src/validation/common/index.ts
@@ -1,18 +1,0 @@
-export * from './param.validation.error.js';
-export * from './param.validation.js';
-export * from './param.validation.with.type.metadata.js';
-
-// builders
-export * from './builders/greater.than.builder.js';
-export * from './builders/is.integer.builder.js';
-export * from './builders/is.number.builder.js';
-export * from './builders/is.string.builder.js';
-export * from './builders/is.valid.instance.builder.js';
-export * from './builders/is.valid.instance.of.builder.js';
-export * from './builders/less.than.builder.js';
-
-// utility
-export * from './utility/builder.utility.js';
-export * from './utility/common.utility.js';
-export * from './utility/decorator.utility.js';
-export * from './utility/error.utility.js';
