--- src/validation/decorators/index.ts
@@ -1,11 +1,0 @@
-// function decorators
-export * from './function/validate.params.decorator.js';
-
-// parameter decorators
-export * from './parameter/greater.than.decorator.js';
-export * from './parameter/is.integer.decorator.js';
-export * from './parameter/is.number.decorator.js';
-export * from './parameter/is.string.decorator.js';
-export * from './parameter/is.valid.instance.decorator.js';
-export * from './parameter/is.valid.instance.of.decorator.js';
-export * from './parameter/less.than.decorator.js';
