+++ src/validation/decorators/parameter/is.number.decorator.ts
@@ -2,7 +2,7 @@ import { buildParamValidationDecorator } from '../../common/utility/decorator.ut
 import { isNumberBuilder } from '../../common/builders/is.number.builder.js';
 
 /**
- * Parameter decorators. Register this validations as needing to be validated as being a number when the method is called.
+ * Parameter decorators. Register this validation as needing to be validated as being a number when the method is called.
  *
  * Example:
  * class MyService {
