+++ src/validation/decorators/parameter/is.string.decorator.ts
@@ -2,7 +2,7 @@ import { buildParamValidationDecorator } from '../../common/utility/decorator.ut
 import { isStringBuilder } from '../../common/builders/is.string.builder.js';
 
 /**
- * Parameter decorators. Register this validations as needing to be validated as being a string when the method is called.
+ * Parameter decorators. Register this validation as needing to be validated as being a string when the method is called.
  *
  * Example:
  * class MyService {
