+++ src/validation/decorators/parameter/is.valid.instance.decorator.ts
@@ -2,7 +2,7 @@ import { buildParamValidationDecorator } from '../../common/utility/decorator.ut
 import { isValidInstanceBuilder } from '../../common/builders/is.valid.instance.builder.js';
 
 /**
- * Parameter decorator. Register this validations as needing to be validated against its specified type when the method is called. It will make sure
+ * Parameter decorator. Register this validation as needing to be validated against its specified type when the method is called. It will make sure
  * the parameter is indeed an instance of the specified type. Additionally, it will validate any property validations applied via class-validator.
  *
  * Note: This parameter decorator can only be used on a parameter whose type is a single class. To validate a parameter whose type is a union of
