+++ src/validation/decorators/parameter/is.valid.instance.of.decorator.ts
@@ -4,7 +4,7 @@ import { isValidInstanceOfBuilder } from '../../common/builders/is.valid.instanc
 import { NonEmptyArray } from '../../../non.empty.array.js';
 
 /**
- * Parameter decorator. Register this validations as needing to be validated against one of several specified types when the method is called. It will
+ * Parameter decorator. Register this validation as needing to be validated against one of several specified types when the method is called. It will
  * make sure the parameter is indeed an instance of one of the specified type. Additionally, it will validate any property validations applied via
  * class-validator.
  *
