+++ src/validation/index.ts
@@ -1,6 +1,42 @@
 export * from './protocol.validation.pipe.js';
 
-// sub-modules
-export * from './common/index.js';
-export * from './decorators/index.js';
-export * from './validations/index.js';
+// common
+export * from './common/param.validation.error.js';
+export * from './common/param.validation.js';
+export * from './common/param.validation.with.type.metadata.js';
+
+// builders
+export * from './common/builders/greater.than.builder.js';
+export * from './common/builders/is.integer.builder.js';
+export * from './common/builders/is.number.builder.js';
+export * from './common/builders/is.string.builder.js';
+export * from './common/builders/is.valid.instance.builder.js';
+export * from './common/builders/is.valid.instance.of.builder.js';
+export * from './common/builders/less.than.builder.js';
+
+// utility
+export * from './common/utility/builder.utility.js';
+export * from './common/utility/common.utility.js';
+export * from './common/utility/decorator.utility.js';
+export * from './common/utility/error.utility.js';
+
+// function decorators
+export * from './decorators/function/validate.params.decorator.js';
+
+// parameter decorators
+export * from './decorators/parameter/greater.than.decorator.js';
+export * from './decorators/parameter/is.integer.decorator.js';
+export * from './decorators/parameter/is.number.decorator.js';
+export * from './decorators/parameter/is.string.decorator.js';
+export * from './decorators/parameter/is.valid.instance.decorator.js';
+export * from './decorators/parameter/is.valid.instance.of.decorator.js';
+export * from './decorators/parameter/less.than.decorator.js';
+
+// Validations
+export * from './validations/greater.than.js';
+export * from './validations/is.integer.js';
+export * from './validations/is.number.js';
+export * from './validations/is.string.js';
+export * from './validations/is.valid.instance.of.js';
+export * from './validations/is.valid.instance.js';
+export * from './validations/less.than.js';
