--- src/validation/validations/index.ts
@@ -1,7 +1,0 @@
-export * from './greater.than.js';
-export * from './is.integer.js';
-export * from './is.number.js';
-export * from './is.string.js';
-export * from './is.valid.instance.of.js';
-export * from './is.valid.instance.js';
-export * from './less.than.js';
