+++ tsconfig.json
@@ -4,7 +4,7 @@
     "moduleResolution": "node",
     "target": "ES2020",
     "lib": ["ES2020"],
-    "declaration": true,
+    "declaration": false,
     "noImplicitAny": false,
     "removeComments": true,
     "noLib": false,
